snd: virtio: Support hardware pointer manipulation from device
With this feature the device is able to update the hardware pointer of
the driver without throughing an interrupt. The interrupt is only used
to wake up a blocking application.
With this solution no additional worker thread is required to support
MMAP audio buffers to applications. Therefore no period time estimation
has to be done. The device provides exactly its period time to the
device and application. This means device and driver running synchrone
with the same period size. This avoids additional buffering. Therefore
smaller audio delays can be achieved.

Change-Id: I127c6e0b3e561be0d7548b3fd238d84e85d75bf5
Signed-off-by: Timo Wischer <twischer@de.adit-jv.com>